+++ predators/predators/Predator.h
@@ -3,7 +3,7 @@
 class Predator : public Fauna
 {
 	public:
-		Predator(float myPosX, float myPosY, float myRadius, int myEnergy, bool mySex,
+		Predator(float myPosX, float myPosY, float myRadius, float myEnergy, bool mySex,
 			int mySpeed, int myAge, int myHungerLevel,
 			float myMetabolicRate, int myLustLevel, int myVisionRange);
