+++ predators/predators/Prey.cpp
@@ -1,7 +1,7 @@
 #include "Prey.h"
 #include "Fauna.h"
 
-Prey::Prey(float myPosX, float myPosY, float myRadius, int myEnergy,bool mySex,
+Prey::Prey(float myPosX, float myPosY, float myRadius, float myEnergy,bool mySex,
 	int mySpeed, int myAge, int myHungerLevel,
 	float myMetabolicRate, int myLustLevel, int myVisionRange) : Fauna(myPosX, myPosY, myRadius, myEnergy,
 		mySex, mySpeed, myAge, myHungerLevel,
