+++ predators/predators/Prey.h
@@ -4,7 +4,7 @@ class Prey : public Fauna
 {
 
 public:
-	Prey(float myPosX, float myPosY, float myRadius, int myEnergy,bool mySex,
+	Prey(float myPosX, float myPosY, float myRadius, float myEnergy,bool mySex,
 		int mySpeed, int myAge, int myHungerLevel,
 		float myMetabolicRate, int myLustLevel, int myVisionRange);
